move tests to attention/
